% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{BOOKTRACKR_REQUIRED_COLUMNS}
\alias{BOOKTRACKR_REQUIRED_COLUMNS}
\title{The columns required in `data.frame` to be used with bookTrackR functions}
\usage{
BOOKTRACKR_REQUIRED_COLUMNS()
}
\description{
The columns required in `data.frame` to be used with bookTrackR functions
}
