% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_book_df}
\alias{validate_book_df}
\title{Function to validate book \code{data.frame} is formatted correctly}
\usage{
validate_book_df(book_df)
}
\arguments{
\item{book_df}{\code{data.frame} to be validated.
Required columns are listed in /code{/link{BOOKTRACKR_REQUIRED_COLUMNS}}}
}
\description{
Function to validate book \code{data.frame} is formatted correctly
}
